added some clenups for the zip file and extracted dir

+++ script.sh
@@ -4,15 +4,21 @@ DATE=$(date "+%Y%m%d%H%M")
 # Log file path
 log_file="/var/log/script.log"
 
-update_file="/opt/servicenow/mid/agent/export/update_xxxx.zip"
+update_file="/opt/servicenow/mid/agent/export/update.zip"
 drass_destination= $(pwd)
 
 echo "Started sync at ${DATE}"  >> "$log_file"
-# Check if the update file exists
-while [ ! -e "$update_file" ]; do
-    echo "Waiting for update file..." >> "$log_file"
-    sleep 5
-done
+
+# # Check if the update file exists
+# while [ ! -e "$update_file" ]; do
+#     echo "Waiting for update file..." >> "$log_file"
+#     sleep 5
+# done
+
+if [ -z "$update_file" ]; then
+    echo "Error: No update file (*.zip) found in the project directory. Please check your folder." >> "$log_file"
+    exit 1
+fi
 
 # Extract the contents of the update file.
 temp_extracted_dir=$(mktemp -d)
@@ -31,8 +37,17 @@ for file in "$temp_extracted_dir"*; do
     fi
 done
 
-# Clean up temporary directory
-rm -r "$temp_extracted_dir"
+# Clean up the extracted update file if it exists
+if [ -d "$temp_extracted_dir" ]; then
+    rm -r "$temp_extracted_dir"
+    echo "Extracted update files deleted." >> "$log_file"
+fi
+
+# Delete the update ZIP file
+if [ -e "$update_file" ]; then
+    rm "$update_file"
+    echo "Update ZIP file deleted." >> "$log_file"
+fi
 
 echo "Update applied successfully." >> "$log_file"
